Fix: properly resolve parameter reference objects
Previously when the parameter provide a $ref it will not get resolved
and thus the schema of the parameter will be properly defined. And also
an error will be raised, wrongfully, for missing referenced objects even
if the referenced object is defined in the OpenAPI spec.

This commit fixes this issue by properly resolving the referenced object
and also raising an error if the referenced object is not found.